added more demos and updated readme
