+++ UICatalog/Scenarios/Buttons.cs
@@ -18,11 +18,12 @@ namespace UICatalog {
 			};
 			top.Add (win);
 
-			var button = new Button (10, 10, "This is a button") { Clicked = () => MessageBox.Query (50, 7, "Message", "Question?", "Yes", "No") };
+			var button = new Button (1, 1, "This is a button") { Clicked = () => MessageBox.Query (30, 7, "Message", "Question?", "Yes", "No") };
 			win.Add (button);
+
+			win.Add (new Button (20, 1, "PRESS ME") { Clicked = () => MessageBox.Query (30, 7, "Message", "Question?", "Yes", "No") });
 
 			Application.Run (top);
 		}
-
 	}
 }
